Rename change password form file name to match component name
