+++ app/pages/settings/ChangePasswordForm.spec.js
@@ -2,7 +2,7 @@
 import React from 'react';
 import { shallow } from 'enzyme';
 import assert from 'assert';
-import ChangePasswordForm from './change-password-form';
+import ChangePasswordForm from './ChangePasswordForm';
 
 describe('ChangePasswordForm', () => {
